Change value of INVALIDATED_TICK_DATA
This value is assigned to tick_data.epoch to signal that this
going to be an empty tick, which must not be overwritten by
incoming BroadcastFutureTickData messages.
The value has been changed to the max. value of the data
type used for epoch to avoid future when increasing MAX_NUMBER_EPOCH.