Add recovery from missmatching txBodyDigest if expectedNextTickTransactionDigest is zero
To avoid beeing stuck on finding a quorum for the current tick on the txBodyDigest if the
expectedNextTickTransactionDigest was changed after voting.

For example:
Node I has all transactions and votes with a valid txBodyDigest
and expectedNextTickTransactionDigest.

Node II does not receive TickData and timesout -> sets
expectednextTickTransactionDigest to zero.

Nodes would be stuck on consensus finding if nodes take into acount txBodyDigest.