Add special command to ignore self-generated computor list
This commit adds a special command to ignore the self-generated
computorlist during the next start. This allows to recover from a
critical situation #2.

On a critical situation #2 send special command with qubic-cli.
Shutdown Qubic properly (escape key) to save systemfile
On the next restart (if systemfile is present) the self-generated
computor list is ignored.