Speed-up TickData construction by tick leader
We now only consider tx for random selection that are scheduled for the
tick we construct the TickData of.

Before this fix, the expensive function random() was called many million
times, because random() was called before checking if a candidate slot
contained a tx scheduled for the tick (or any tx at all).