Add plotform header file to surpress compiler warnings (#399)
* Add plotform header file to surpress compiler warnings

Header allows to surpress compiler specific warnings without yielding warnings/errors in other compilers.

Example:
We need to silent some warning in the clang build. This can be achieved with a `#pragma clang ...`. However this would yield warning when buidling with msvc since this pragma is not known. This header allows to suprress the warnings for clang without the need for guards around every pragma.