Fix data types in m256 tests
Problem with higher warning levels MSVC points to a converstion problem between int and unsigned char and potential loss of data. No actuall loss happened due to used values but with a `int8_t` the correct value is used.
Additionally one of the tests initializes a m256 with 4 negative values but functions expects `unsigned long long`