Fix error in refresh peers logic
Nodes regurlarly close random connected peers. However the logic
required publicPees to be exchanged before it potentially is selected
for closing. It is to expect that a lot of incoming request do ignore
the exchangePublicPeer process and hence are never valid for selection.
This is inparticullarly critical if the peer is not closing the
connection properly from his side leading to slowly filling all
connection slots.