Clang port: m265i test cases (#414)
* Update README_CLANG.md

* Unify required CMake version across all projects

* Add compiler_warnings to vcxproj file

* Adjust type in m256.h to follow qstdint.h

* Fix call to _rdrand64_step

Clang is stricter regarding type aliasing rules and will emit warnings or errors if it detects potentially unsafe aliasing unless told otherwise. This commit explicitly reinterpret_casts the arguments to `_rdrand64_step` to `unsigned long long*`

* Remove obsolete volatile specifier

* Add static cast to avoid type error during comparison

* Fix error due to stricter clang rules regarding volatile

* clang m256: Fix type mismatch on comparisons

Clang is stricter with type mismatches during comparison. This commit fixes the types by explicitly cast to correct type.

* Adjust m256.h to work with clang

Clang is more strict with resprect to alignment of types. Hence Some
changes are needed that the m256.h header is working with it.
Inparticualar these are the following changes:

- Introduced warning suppression for cast alignment and self-assignment.
- Updated m256i_intr() to return by value, added volatile overload,
  and introduced new set_intrinsic_value() setter.
- Changed __m256i_convert functions to return by value.
- Adjusted test suite for new practices and warning suppression.

* Clang: Remove supression of warnings in m256

* Adress reviwer comments

* Add Benchmark for m256i running in EFI

* Fix various cmake errors

* CMake: Fix qubic.efi build

* Add Benchmark to Visual Studio Project

* Adjust README_CLANG.md