Change typedef of CHAR16 to wchar_t
The type of string literals in c++ is wchar_t. However Windows does not
follow the standard and uses a 16 bit type for it. Other compiler follow
the standard and use 32 bit type. The behaviour of clang can be changes
with the flag -f-short-wchar however the types wchar_t and unsigned char
(CHAR16) are fundamentally different and no implicit conversion is
possible.

To be able to compile on both system we redefine CHAR16 as wchar_t.
Violating the UEFI Specification. However due to the same representation
this should not further lead to problems.