Fix error "unresolved symbol memcpy"
For some types of structs, MSVC generates calls to memcpy on assignment.
This is a problem, because we don't link with the runtime lib in UEFI,
which defines memcpy.

This commit fixes the issue for QPI::HashMap values, because the issue
was encountered there during the development of the contract NOSTROMO.