Fix error "unresolved symbol memcpy"
For QPI::Array of some types of structs, MSVC generates calls to memcpy
on assignment. This is a problem, because we don't link with the
runtime lib in UEFI, which defines memcpy.

This commit fixes the issue by defining the copy constructor and the
assignment operator for QPI::Array.

The issue was encountered during the development of the contract
NOSTROMO.