Decouples TARGET_TICK_DURATION and tick_storage allocation
This commit decouples the allocation of the tick_storage, tx_addon and
logging by introducing a new `#define TICK_DURATION_FOR_ALLOCATION_MS`
that inidicated the assumed tick duration in miliseconds.

To be able to use sub second tick durations the computation of
MAX_NUMBER_OF_TICKS computation is adjusted to be based on total
number of miliseconds in a epoch.